requirements.txt
aiohttp==3.8.1
aiohttp_cors==0.7.0
aioredis==2.0.1
aioretry==5.0.2
autopep8==1.6.0
dependency-injector==4.39.1
psycopg2==2.9.3
python-decouple==3.6
python-engineio==4.3.2
python-socketio==5.6.0
sqlalchemy==1.4.36